Fix and simplify Whipe script

+++ Labs/whipe.py
@@ -4,25 +4,29 @@
 import uos as os 
 
 def whipe_dir( path=".",sub=True):
+        print( "Whipe path {}".format(path) )
         l = os.listdir(path)
         l.sort()
-        for f in l:
-            child = "%s/%s" % (path, f)
-            st = os.stat(child)
-            if st[0] & 0x4000:  # stat.S_IFDIR
-                print("   <dir> %s" % child)
-                if sub:
-                    
-                    whipe_dir(child,sub)
-                    os.rmdir(child)
-            else:
-                os.remove(child)
+        #print(l)
+        if l != ['']:
+            for f in l:
+                child = "{}/{}".format(path, f)
+                #print(" - "+child)
+                st = os.stat(child)
+                if st[0] & 0x4000:  # stat.S_IFDIR
+                    if sub:
+                        whipe_dir(child,sub)
+                        try:
+                            os.rmdir(child)
+                        except:
+                            print("Error deleting folder {}".format(child))
+                else: # File
+                    try:  
+                        os.remove(child)
+                    except:
+                        print("Error deleting file {}".format(child))
 
-def main():
-    a = input("Do you wish to whipe the /flash folder and all underlying files ?")
-    if a=='Y' or a=='y':
-        whipe_dir(path='/flash',sub=True)
+whipe_dir(path='/flash')
 
-if __name__ == "__main__":
-    main()
+
